update reports font and order

+++ database/migrations/2025_11_07_061115_add_tid_in_funds_table_for_transaction.php
@@ -12,10 +12,12 @@ return new class extends Migration
    public function up(): void
 {
     Schema::table('funds', function (Blueprint $table) {
-        $table->foreignId('tid')
-              ->nullable()
-              ->constrained('transactions')
-              ->onDelete('set null'); // optional
+        if (!Schema::hasColumn('funds', 'tid')) {
+            $table->foreignId('tid')
+                  ->nullable()
+                  ->constrained('transactions')
+                  ->onDelete('set null'); // optional
+        }
     });
 }
